@@ -8,18 +8,22 @@ import { MenuComponent } from './restaurant-detail/menu/menu.component';
 import { ReviewsComponent } from './restaurant-detail/reviews/reviews.component';
 import { OrderComponent } from './order/order.component';
 import { OrderSummaryComponent } from './order-summary/order-summary.component';
+import { LoginComponent } from './security/login/login.component';
+import { LoggedInGuard } from './security/loggedin.guard';
 
 export const ROUTES: Routes = [
   { path: '', component: HomeComponent },
-  { path: 'restaurants', component: RestaurantsComponent },
+  { path: 'login/:to', component: LoginComponent },
+  { path: 'login', component: LoginComponent },
   { path: 'restaurants/:id', component: RestaurantDetailComponent,
-  children: [
-    { path: '', redirectTo: 'menu', pathMatch: 'full'},
-    { path: 'menu', component: MenuComponent},
-    { path: 'reviews', component: ReviewsComponent}
-  ]
-  },
-  { path: 'order', loadChildren: './order/order.module#OrderModule' },
+    children: [
+      { path: '', redirectTo: 'menu', pathMatch: 'full'},
+      { path: 'menu', component: MenuComponent},
+      { path: 'reviews', component: ReviewsComponent}]
+    },
+  { path: 'restaurants', component: RestaurantsComponent },
+  { path: 'order', loadChildren: './order/order.module#OrderModule',
+    canLoad: [LoggedInGuard], canActivate: [LoggedInGuard]},
   { path: 'order-summary', component: OrderSummaryComponent },
-  { path: 'about', loadChildren: './about/about.module#AboutModule' }
+  { path: 'about', loadChildren: './about/about.module#AboutModule'}
 ];
