@@ -85,7 +85,7 @@ export class OrderComponent implements OnInit {
       .map((item:CartItem) => new OrderItem(item.quantity, item.menuItem.id))
     
     this.orderService.checkOrder(order)
-      .subscribe((orderId: Order) => {
+      .subscribe((orderId: string) => {
         this.router.navigate(['/order-summary'])
         this.orderService.clear()
       })
