@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
-import { Http, Headers, RequestOptions } from '@angular/http';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 import 'rxjs/add/operator/map';
 
 
@@ -10,12 +10,15 @@ import { Order, OrderItem } from './order.model';
 
 
 import { MEAT_API } from '../app.api';
+import { LoginService } from '../security/login/login.service';
 
 
 
 @Injectable()
 export class OrderService {
-  constructor(private cartService: ShoppingCartService, private http: Http){}
+  constructor(private cartService: ShoppingCartService,
+              private http: HttpClient,
+              private loginService: LoginService){}
 
   itensValue(): number {
     return this.cartService.total();
@@ -41,13 +44,12 @@ export class OrderService {
     this.cartService.clear();
   }
 
-  checkOrder(order: Order): Observable<Order> {
-    const headers = new Headers();
-    headers.append('Content-Type', 'application/json');
-    return this.http.post(`${MEAT_API}/orders`, 
-                          JSON.stringify(order), 
-                          new RequestOptions({headers: headers}))
-                    .map(response => response.json())
+  checkOrder(order: Order): Observable<string> {
+    let headers = new HttpHeaders()
+    if(this.loginService.isLoggedIn()) {
+      headers = headers.set('Authorization', `Bearer ${this.loginService.user.accessToken}`)
+    }
+    return this.http.post<Order>(`${MEAT_API}/orders`, order, {headers: headers})
                     .map(order => order.id)
   }
 
